requirements.txt
alpaca-py==0.42.0
crispy-bootstrap5==2025.6
django==5.2.1
django-cors-headers==4.7.0
django-crispy-forms==2.4
django-dotenv==1.4.2
djangorestframework==3.16.0
pandas_market_calendars==5.1.1
psycopg==3.2.9
psycopg2-binary==2.9.10
tzdata==2025.2
yfinance==0.2.64
